requirements.txt
alembic==1.13.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Dance==7.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
mysql-connector-python==9.0.0
oauthlib==3.2.2
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
URLObject==2.4.3
Werkzeug==3.0.4
zipp==3.20.2
